+++ tico/quantization/wrapq/examples/quantize_full_qmodel_with_gptq.py
@@ -75,8 +75,6 @@ DATASET_CONFIG = "wikitext-2-raw-v1"
 TRAIN_SPLIT = "train"
 TEST_SPLIT = "test"
 
-left_pad = False#True
-
 # -------------------------------------------------------------------------
 # Helper — copy GPTQ (scale, zp) into PTQ observers
 # -------------------------------------------------------------------------
@@ -113,137 +111,7 @@ def inject_gptq_qparams(
             continue
 
         assert isinstance(obs, AffineObserverBase)
-        obs.load_qparams(quantizer.scale, quantizer.zero, lock=True)
-
-def pad_input(input, pad_token, max_seq_len, right: bool = True):
-    """Pad a tensor to a maximum sequence length using the specified pad token."""
-    pads = torch.full(
-        (input.shape[0], max_seq_len - input.shape[1]),
-        fill_value=pad_token,
-        device=input.device,
-    )
-    if right is True:
-        res = torch.cat((input, pads), dim=1)
-    else:
-        res = torch.cat((pads, input), dim=1)
-
-    return res
-class PrefillDecodeUtils:
-    def __init__(self, max_seq_len, config, device):
-        self.max_seq_len = max_seq_len
-        self.device = device
-        self.pos_embeds = _build_rope_templates_from_config(
-            config, max_seq=max_seq_len, device=device, dtype=torch.float32
-        )
-
-    def build_attention_mask_for_padded_input(self, cur_seq_len, right_padding = False, mask_value:float = -120.0, prefill: bool = True, pad_mask = None):
-        dtype = torch.float32
-      #  mask = torch.full((1, max_seq_len, max_seq_len), mask_value, device=device, dtype=dtype)
-      #  
-      #  #return mask
-      #  if right_padding:
-      #      for i in range(max_seq_len):
-      #          for j in range(max_seq_len):
-      #              if i >= j and j < cur_seq_len:
-      #                  mask[..., i, j] = 0
-      #  else:
-      #      for i in range(max_seq_len):
-      #          for j in range(max_seq_len):
-      #              if i >= j and j >= max_seq_len - cur_seq_len:
-      #                  mask[..., i, j] = 0
-
-        max_seq_len = self.max_seq_len -  1 if prefill is True else self.max_seq_len
-        if pad_mask is None:
-            pad_mask = [i < cur_seq_len for i in range(max_seq_len)] if right_padding else [i >= max_seq_len - cur_seq_len for i in range(max_seq_len)]
-            pad_mask = torch.tensor(pad_mask, dtype = torch.bool, device = self.device)
-        causal_mask = torch.full((1, max_seq_len, max_seq_len), 1, device=self.device, dtype=dtype)
-        causal_mask = torch.triu(causal_mask, diagonal=1)
-        causal_mask = causal_mask == 0 # negating
-        res_mask = torch.logical_and(pad_mask, causal_mask)
-        mask_res = torch.zeros((1, max_seq_len, max_seq_len), device=self.device, dtype=dtype)
-        mask_res = mask_res.masked_fill(~res_mask, mask_value)
-        return mask_res
-        #assert torch.equal(mask_res, mask)
-        
-        #return mask
-    
-    def build_position_embeddings_for_padded_input(self, position_embeddings, cur_seq_len, position_ids, right_padding = False, prefill: bool = True):
-        dtype = torch.float32
-        #position_embeddings = self.prefill_model.wrapped.model.wrapped.get_position_embeddings_for(dtype, self.device)
-        #max_seq_len = self.max_seq_len -  1 if prefill is True else self.max_seq_len
-        #cos = torch.ones_like(position_embeddings[0][:, : max_seq_len, :])
-        #sin = torch.zeros_like(position_embeddings[1][:, : max_seq_len, :])
-        #
-        #sl_cos, sl_sin = position_embeddings    
-        #sl_cos = sl_cos[:, : cur_seq_len, :]
-        #sl_sin = sl_sin[:, : cur_seq_len, :]
-        #if right_padding is True:
-        #    cos[..., :cur_seq_len, :] = sl_cos
-        #    sin[..., :cur_seq_len, :] = sl_sin
-        #else:
-        #    # left padding
-        #    cos[..., max_seq_len - cur_seq_len : max_seq_len, :] = sl_cos
-        #    sin[..., max_seq_len - cur_seq_len : max_seq_len, :] = sl_sin
-        #    
-        #return (cos, sin)
-        
-        sl_cos, sl_sin = position_embeddings
-        sl_cos = sl_cos[..., :-1, :]
-        sl_sin = sl_sin[..., :-1, :]
-        cos = sl_cos[..., position_ids.to(sl_cos.device).squeeze(0), :]
-        sin = sl_sin[..., position_ids.to(sl_sin.device).squeeze(0), :]
-        
-        #cos = torch.gather(sl_cos, index = position_ids.to(sl_cos.device).squeeze(0).unsqueeze(1))
-        #sin = torch.gather(sl_sin, index = position_ids.to(sl_sin.device).squeeze(0).unsqueeze(1))
-        return (cos, sin)
-    
-    def get_input_for_decode_model(self, input, past_key_values, cur_seq_len, right_padding=False, pad_mask = None):
-        dtype = torch.float32
-        next_token = torch.tensor([[input[..., -1]]], device = input.device)
-        if right_padding:
-          #  attention_mask = _build_decode_attention_mask(
-          #      batch_size=1,
-          #      past_len=cur_seq_len,
-          #      max_seq=self.max_seq_len,
-          #      device=self.device,
-          #      dtype=dtype,
-          #  )
-##
-          #  position_embeddings = _slice_rope(
-          #      self.pos_embeds[0],  # cos
-          #      self.pos_embeds[1],  # sin
-          #      position=cur_seq_len,
-          #      batch_size=1,
-          #      device=self.device,
-          #      dtype=dtype,
-          #  )
-            attention_mask = self.build_attention_mask_for_padded_input(cur_seq_len, right_padding=True, prefill=False, pad_mask = pad_mask)
-            attention_mask = attention_mask[..., -1, :].unsqueeze(0) # last row
-            #position_embeddings = self.build_position_embeddings_for_padded_input(self.pos_embeds, cur_seq_len, right_padding=True, prefill=False)
-            #(cos, sin) = position_embeddings
-            #cos = cos[..., -1, :].unsqueeze(0)
-            #sin = sin[..., -1, :].unsqueeze(0)
-            #position_embeddings = (cos, sin)
-            position_embeddings = self.build_position_embeddings_for_padded_input(self.pos_embeds, cur_seq_len, position_ids=torch.tensor([[cur_seq_len - 1]]), right_padding=True, prefill=False)
-        else:
-            attention_mask = self.build_attention_mask_for_padded_input(cur_seq_len, right_padding=False, prefill=False, pad_mask = pad_mask)
-            attention_mask = attention_mask[..., -1, :].unsqueeze(0) # last row
-            #position_embeddings = self.build_position_embeddings_for_padded_input(self.pos_embeds, cur_seq_len, position_ids=torch.tensor([[cur_seq_len]]), right_padding=False, prefill=False )
-            #(cos, sin) = position_embeddings
-            #cos = cos[..., -1, :].unsqueeze(0)
-            #sin = sin[..., -1, :].unsqueeze(0)
-            #position_embeddings = (cos, sin)
-            position_embeddings = self.build_position_embeddings_for_padded_input(self.pos_embeds, cur_seq_len, position_ids=torch.tensor([[cur_seq_len - 1]]), right_padding=False, prefill=False)
-
-        # fill in input
-        inputs = {}
-        inputs["input_ids"] = next_token
-        inputs["attention_mask"] = attention_mask
-        inputs["position_embeddings"] = position_embeddings
-        inputs["past_key_values"] = past_key_values
-        inputs["use_cache"] = True
-
-        return inputs
+        obs.load_qparams(quantizer.scale, quantizer.zero, lock=True)    
         
 def get_decode_input(
     prefill_model,
@@ -261,11 +129,11 @@ def get_decode_input(
     prefill_max_seq_len = max_seq_len - 1
     prefill_input = pad_input(prefill_input, pad_token_id, prefill_max_seq_len, right = not left_pad)
     prefill_attn_mask = (prefill_input != pad_token_id).to(device)
-    attn_mask = helper.build_attention_mask_for_padded_input(prefill_seq_len, right_padding=not left_pad, pad_mask = prefill_attn_mask)
+    attn_mask = helper.build_attention_mask_for_padded_input(prefill_attn_mask, prefill=True)
     
     prefill_position_ids = prefill_attn_mask.long().cumsum(-1) - 1
     prefill_position_ids.masked_fill_(prefill_attn_mask == 0, 1)
-    position_embeddings = helper.build_position_embeddings_for_padded_input(ropes, prefill_seq_len, prefill_position_ids, right_padding=not left_pad)
+    position_embeddings = helper.build_position_embeddings_for_padded_input(ropes, prefill_position_ids)
     
     with torch.no_grad():
         # run prefill model to get kv-cache
@@ -276,130 +144,9 @@ def get_decode_input(
     prefill_input = torch.concat([prefill_input, next_token], dim=1)
     attn_mask = (prefill_input != pad_token_id).to(device)
     
-    dec_inputs = helper.get_input_for_decode_model(prefill_input, past_key_values=outputs.past_key_values, cur_seq_len = prefill_seq_len + 1, right_padding=not left_pad, pad_mask = attn_mask)
+    dec_inputs = helper.get_input_for_decode_model(prefill_input, past_key_values=outputs.past_key_values, pad_mask = attn_mask, cur_seq_len = prefill_seq_len + 1)
     return dec_inputs
-    #dtype=torch.float32
-    #attention_mask = _build_decode_attention_mask(
-    #    batch_size=1,
-    #    past_len=prefill_seq_len,
-    #    max_seq=max_seq_len,
-    #    device=device,
-    #    dtype=dtype,
-    #)
-#
-    #rope_cos, rope_sin = ropes
-    #position_embeddings = _slice_rope(
-    #    rope_cos,
-    #    rope_sin,
-    #    position=prefill_seq_len - 1,
-    #    batch_size=1,
-    #    device=device,
-    #    dtype=dtype,
-    #)
-
-    # fill in input
-    inputs = {}
-    inputs["input_ids"] = torch.tensor([[next_token]])
-    inputs["attention_mask"] = attention_mask
-    inputs["position_embeddings"] = position_embeddings
-    inputs["past_key_values"] = outputs.past_key_values
-    return inputs
-
-def evaluate_ppl_of_prefill_decode_model_on_dataset(
-    model,
-    dataset,
-    pad_token_id,
-    max_seq_len,
-    seed=0,
-    device: str = "cuda",
-):
-    """Compute perplexity for the prefill-decode logic."""
-
-    config = (
-        model.config
-        if hasattr(model, "config")
-        else model.wrapped.config
-    )
-    rope_cos, rope_sin = _build_rope_templates_from_config(
-        config, max_seq=max_seq_len, device=device, dtype=torch.float32
-    )
-
     
-    if hasattr(model, "device") and model.device.type != device.type:
-        if hasattr(model, "to"):
-            model.to(device)
-
-    torch.manual_seed(seed)
-    nlls = []
-    helper = PrefillDecodeUtils(max_seq_len, config, device)
-    
-    with torch.no_grad():
-        for batch in tqdm.tqdm(dataset):
-            if isinstance(batch, torch.Tensor):
-
-                prefill_seq_len = (
-                    torch.randint(3, max_seq_len - 1, (1,)).cpu().item()
-                ) # max_seq_len - 1# cropped input length
-                prefill_input = batch[..., :prefill_seq_len]  # cropped input
-
-                last_token = batch[..., prefill_seq_len].cpu().unsqueeze(0)
-                #torch.tensor(
-                #    [[torch.argmax(ref_output.logits[:, -1, :], dim=-1).cpu()]]
-                #)
-                if hasattr(model, "wrapped"):
-                    inputs = get_decode_input(
-                        model,
-                        prefill_input,
-                        pad_token_id,
-                        (rope_cos, rope_sin),
-                        max_seq_len,
-                        device,
-                        helper
-                    )
-                    inputs = move_to_device(inputs, device)
-                    output = model(**inputs)
-                else:
-                    input = pad_input(
-                        prefill_input[..., :-1], pad_token_id, max_seq_len - 1, right = not left_pad
-                    )
-                    prefill_attn_mask = input != pad_token_id
-                    prefill_position_ids = prefill_attn_mask.long().cumsum(-1) - 1
-                    prefill_position_ids.masked_fill_(prefill_attn_mask == 0, 1)
-                    prefill_ouput = model(
-                        input.to(model.device), 
-                        attention_mask = prefill_attn_mask.to(model.device), 
-                        position_ids = prefill_position_ids.to(model.device), 
-                        use_cache=True
-                    )
-                    next_token = prefill_input[..., -1:]
-                    decode_attention_mask = torch.ones((1, max_seq_len))
-                    decode_attention_mask[..., :input.shape[-1]] = input != pad_token_id
-                    decode_position_ids = torch.tensor([[prefill_seq_len-1]])
-                    
-                    output = model(
-                        next_token.to(model.device),
-                        past_key_values=prefill_ouput.past_key_values,
-                        attention_mask = decode_attention_mask.to(model.device),
-                        position_ids = decode_position_ids.to(model.device),
-                        use_cache=True,
-                    )
-
-            lm_logits = output.logits
-
-            if torch.isfinite(lm_logits).all():
-                labels = last_token[0].to(device)
-                loss_fct = torch.nn.CrossEntropyLoss(reduction="none")
-                loss = loss_fct(
-                    lm_logits.reshape(-1, lm_logits.size(-1)),
-                    labels.view(-1),
-                )
-                nlls.append(loss)
-
-            torch.cuda.empty_cache()
-    ppl = np.exp(torch.cat(nlls, dim=-1).mean().item())
-    return ppl
-
-
 def save_model_to(q_m, calib_inputs, args, prefill = True, kwargs=None):
     """ Save the whole model in circle format  """
     q_m.eval()
@@ -790,152 +537,8 @@ class PrefillQModelProcessor(QModelProcessor):
             save_model_to(model, calib_inputs, self.args)
 
 
-class PrefillDecodeQModelProcessor(QModelProcessor):
-    """
-    Processor for Prefill-Decode models.
-    Prefill-model computes kv-cache for the user input then each new token is produced by decode-model wit upadted kv-cache.
-    """
-
-    def __init__(self, model, tokenizer, args):
-        """Initialize the prefill‑decode processor, handling tokenizer pad token and preparing rotary embeddings."""
-        super().__init__(model, tokenizer, args)
-        if tokenizer.pad_token is None:
-            print(
-                "Warning: tokenizer doesn't have pad_token. Prefill-decoding scheme may fail."
-            )
-            tokenizer.pad_token = tokenizer.eos_token
-
-        rope_cos, rope_sin = _build_rope_templates_from_config(
-            self.model.config,
-            max_seq=self.args.max_seq_len,
-            device=self.device,
-            dtype=torch.float32,
-        )
-        self.rope_cos = rope_cos
-        self.rope_sin = rope_sin
-
-        # debug padding
-
-        #    inputs = tokenizer("Hello!", return_tensors="pt", max_length=args.max_seq_len - 1, padding='max_length', padding_side="right").input_ids.to(device)
-        #    #inputs = tokenizer("Hello! How are you?", return_tensors="pt").input_ids.to(device)
-        #    model.config.use_cache = True
-        #    model.config._attn_implementation = "eager"
-        #    out_ids = model.generate(inputs)
-        #    output = tokenizer.decode(out_ids.squeeze(), skip_special_tokens=True)
-
-    #        prefill_decode_ppl = evaluate_ppl_of_prefill_decode_model_on_dataset(q_m, q_m, calib_inputs, pad_token_id=tokenizer.pad_token_id, max_seq_len = args.max_seq_len, seed = args.seed, device = device)
-    #
-    #        print("\n┌── Wikitext-2 prefill_decode initial calibration perplexity──")
-    #        print(f"│ FP32 : {prefill_decode_ppl:8.2f}")
-    #        print("└───────────────────────────────────────────")
-
-    def run_ptq(self, q_m, calib_inputs):
-        """Run PTQ for the prefill‑decode pipeline, calibrating for padding and kv-cache."""
-        
-        # ?? pre_ptq_model = copy.deepcopy(q_m).to("cpu")  # to be used in decode quntizing
-
-        # get prefill_model
-        q_m = super()._run_ptq(q_m, calib_inputs)
-
-        prefill_decode_ppl = evaluate_ppl_of_prefill_decode_model_on_dataset(
-            q_m,
-            calib_inputs,
-            pad_token_id=self.tokenizer.pad_token_id,
-            max_seq_len=self.args.max_seq_len,
-            seed=self.args.seed,
-            device=self.device,
-        )
-
-        print(
-            "\n┌── Wikitext-2 prefill_prefill train calibration perplexity ─────────────"
-        )
-        print(f"│ int16 : {prefill_decode_ppl:8.2f}")
-        print("└───────────────────────────────────────────")
-
-        torch.manual_seed(self.args.seed)
-
-        return q_m
-
-    def evaluate_original(self, dataset_test):
-        """Evaluate the original (FP) model using the prefill‑decode pipeline."""
-        super().evaluate_original(dataset_test)
-
-        test_inputs = self.get_tokenized_inputs(dataset_test, shuffle=False)
-        prefill_decode_ppl = evaluate_ppl_of_prefill_decode_model_on_dataset(
-            self.model,
-            test_inputs,
-            pad_token_id=self.tokenizer.pad_token_id,
-            max_seq_len=self.args.max_seq_len,
-            seed=self.args.seed,
-            device=self.device,
-        )
-
-        print("\n┌── Wikitext-2 prefill_prefill original test perplexity ─────────────")
-        print(f"│ FP32 : {prefill_decode_ppl:8.2f}")
-        print("└───────────────────────────────────────────")
-
-    def evaluate_quantized(self, model, dataset_test):
-        """Evaluate the quantized prefill‑decode model on the test dataset."""
-        
-        evaluate(model, self.tokenizer, dataset_test, self.args, quantized=True)
-
-        test_inputs = self.get_tokenized_inputs(dataset_test, shuffle=False)
-        prefill_decode_ppl = evaluate_ppl_of_prefill_decode_model_on_dataset(
-            model,
-            test_inputs,
-            pad_token_id=self.tokenizer.pad_token_id,
-            max_seq_len=self.args.max_seq_len,
-            seed=self.args.seed,
-            device=self.device,
-        )
-        print("\n┌── Wikitext-2 prefill_decode quantized test perplexity ─────────────")
-        print(f"│ int16 : {prefill_decode_ppl:8.2f}")
-        print("└───────────────────────────────────────────")
-
-    def save_quantized(self, model, calib_inputs):
-        """Save the quantized prefill and decode models (and optionally their layers) to disk."""
-        
-        model.wrapped.config.use_cache = True
-        if self.args.output_dir is not None and "circle_per_layer" in self.args.save:
-            save_layers_to(model, self.args, prefill=True)
-            save_layers_to(model, self.args, prefill=False)
-
-        if self.args.output_dir is not None and "circle_full" in self.args.save:
-            model = model.to("cpu")
-            calib_inputs = list(
-                torch.stack(calib_inputs).reshape(-1, 1, self.args.max_seq_len)
-            )
-            # save prefill model
-            save_model_to(model, [calib_inputs[0][..., :-1]], self.args, prefill=True)  # seq_len = max_seq_len - 1
-
-            # compute example input
-            prefill_seq_len = (
-                torch.randint(3, self.args.max_seq_len - 1, (1,)).cpu().item()
-            )  # cropped input length
-            prefill_input = calib_inputs[0][..., :prefill_seq_len].to(
-                "cpu"
-            )  # cropped input
-
-            inputs = get_decode_input(
-                model,
-                prefill_input,
-                self.tokenizer.pad_token_id,
-                (self.rope_cos.cpu(), self.rope_sin.cpu()),
-                self.args.max_seq_len,
-                "cpu",
-                helper=PrefillDecodeUtils(self.args.max_seq_len, model.wrapped.config, "cpu")
-            )
-            
-            assert "input_ids" in inputs
-            input = inputs.pop("input_ids")
-
-            # save decode model
-            save_model_to(model, [input], self.args, prefill=False, kwargs = inputs) # seq_len = 1
-
 
 def get_qmodel_processor(model, tokenizer, args):
-    if args.prefill_decode:
-        return PrefillDecodeQModelProcessor(model, tokenizer, args)
 
     return PrefillQModelProcessor(model, tokenizer, args)
