+++ tico/quantization/wrapq/wrappers/llama/quant_model.py
@@ -551,10 +551,11 @@ class QuantLlamaModel(QuantModuleBase):
             attention_mask,
             past_len=past_seen_tokens,
         )
-        position_embeddings = self.get_position_embeddings_for(
-            hidden_states,
-            start=past_seen_tokens,
-        )
+        if position_embeddings is None:
+            position_embeddings = self.get_position_embeddings_for(
+                hidden_states,
+                start=past_seen_tokens,
+            )
 
         # decoder layers
         all_hidden_states = () if output_hidden_states else None
